backend/requirements.txt
# API
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
geoalchemy2==0.17.0

# ETL / Data
pandas==2.2.3
geopandas==1.0.1
requests==2.32.3
aiohttp==3.11.11
openpyxl==3.1.5
lxml==5.3.0

# Config
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1

# Auth JWT (Fase 7)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Rate limiting (Fase 7)
slowapi==0.1.9

# ML — predicción de riesgo de desahucio (Fase 7)
scikit-learn==1.6.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1
